# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FreshTrack API
  version: 2.0.0
  description: |
    Enterprise Inventory & Expiration Date Management System API
    
    ## Authentication
    All endpoints except `/auth/login` and `/auth/register` require Bearer token authentication.
    
    ## Multi-tenancy
    Data is isolated by hotel_id. Users can only access data from their assigned hotel.
    
    ## Permissions
    Role-based access control (RBAC) with granular permissions per resource and action.
  contact:
    name: FreshTrack Support
    email: support@freshtrack.app
  license:
    name: Proprietary
servers:
  - url: https://api.freshtrack.app
    description: Production
  - url: http://localhost:3001
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        unit:
          type: string
          enum: [шт, кг, л, упак, порц]
        category_id:
          type: integer
        department_id:
          type: integer
        min_quantity:
          type: number
        is_active:
          type: boolean
        image_url:
          type: string
          nullable: true
    Batch:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        batch_code:
          type: string
        quantity:
          type: number
        expiry_date:
          type: string
          format: date
        production_date:
          type: string
          format: date
          nullable: true
        supplier:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, collected, written_off, transferred]
        expiryStatus:
          type: string
          enum: [expired, today, critical, warning, good]
        daysLeft:
          type: integer
        statusColor:
          type: string
        statusText:
          type: string
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
security:
  - bearerAuth: []
paths:
...
  /api/products:
    get:
      tags: [Inventory]
      summary: List products
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'

    post:
      tags: [Inventory]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, unit, category_id]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                sku:
                  type: string
                unit:
                  type: string
                  enum: [шт, кг, л, упак, порц]
                category_id:
                  type: integer
                min_quantity:
                  type: number
                  minimum: 0
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    $ref: '#/components/schemas/Product'
  /api/batches:
    get:
      tags: [Inventory]
      summary: List batches
      parameters:
        - name: product_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, collected, written_off, transferred, expiring, expired, critical, warning]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: List of batches with expiry data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Batch'

    post:
      tags: [Inventory]
      summary: Create batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, quantity, expiry_date]
              properties:
                product_id:
                  type: integer
                quantity:
                  type: number
                  minimum: 0.001
                expiry_date:
                  type: string
                  format: date
                production_date:
                  type: string
                  format: date
                batch_code:
                  type: string
                supplier:
                  type: string
      responses:
        '201':
          description: Batch created
...